add compose

+++ building/scripts/kalibr_entrypoint.sh
@@ -2,30 +2,46 @@
 
 set -eox pipefail
 
+# NOTE(Jack): These three lines are the classic lines found in the Kalibr dockerfile entrypoint
 export KALIBR_MANUAL_FOCAL_LENGTH_INIT=1
 source "${WORKSPACE}/devel/setup.bash"
 cd "${WORKSPACE}"
 
-"$@"
-command_status=$?
+dataset_specification_json="/temporary/config/dataset_specification.json"
+target_config="/temporary/config/kalibr/april_6x6_80x80cm.yaml"
 
-# NOTE(Jack): Normally an entrypoint script would end with "$@, but we need to do some post processing inside of the
-# running docker container to get the data sorted and moved into the right place. After the command to run kalibr is
-# finished executing here the script below will execute.to move and organize the output diagnostics.
+[[ -f "${dataset_specification_json}" ]] || { echo "Error: json dataset specification does not exist: ${dataset_specification_json}" >&2; exit 1; }
+[[ -f "${target_config}" ]] || { echo "Error: Kalibr target configuration does not exist: ${target_config}" >&2; exit 1; }
 
-# NOTE(Jack): We need to clean the camera name from having slashes (ex. like in a ROS topic) because otherwise the
-# mkdir command below could interpret that as a multilevel path.
-camera_name="${CAMERA_TOPIC#/}"
-camera_name="${camera_name//\//_}"
+# NOTE(Jack): Iteration logic adopted from https://stackoverflow.com/questions/68121082/how-to-iterate-over-json-array-with-jq
+while read bag_i; do
+    # NOTE(Jack): This removes the json quotes
+    bag_i=$(echo "${bag_i}" | jq -r ".")
 
-# TODO(Jack): We should protect against the case where there are no output files to move, if that is an error condition
-# that we realistically need to worry about.
-# TODO(Jack): Define a variable for the data path?
-# TODO(Jack): It might be a smart idea to be more specific about which files we move, for example all files that do not
-# end in .bag or all files that match the kalibr output format pattern, etc.
-# NOTE(Jack): Kalibr does not let us specify an output path or filename for the output diagnostics. Therefore
-# we need to manually move all the outputs into a batch specific (i.e. camera name and model) directory.
-mkdir --parents "/data/kalibr/${camera_name}"
-mv -- /data/*.pdf /data/*.txt /data/*.yaml "/data/kalibr/${camera_name}"
+    while read camera_i; do
+        camera_i=$(echo "${camera_i}" | jq -r ".")
 
-exit "${command_status}"
+        rosrun kalibr kalibr_calibrate_cameras \
+          --bag "/data/${bag_i}" \
+          --dont-show-report \
+          --models ds-none \
+          --target "/temporary${target_config}" \
+          --topics "${camera_i}"
+
+        # NOTE(Jack): We need to clean the camera name from having slashes (ex. like in a ROS topic) because otherwise the
+        # mkdir command below could interpret that as a multilevel path.
+        camera_name="${CAMERA_TOPIC#/}"
+        camera_name="${camera_name//\//_}"
+
+        # TODO(Jack): We should protect against the case where there are no output files to move, if that is an error condition
+        # that we realistically need to worry about.
+        # TODO(Jack): Define a variable for the data path?
+        # TODO(Jack): It might be a smart idea to be more specific about which files we move, for example all files that do not
+        # end in .bag or all files that match the kalibr output format pattern, etc.
+        # NOTE(Jack): Kalibr does not let us specify an output path or filename for the output diagnostics. Therefore
+        # we need to manually move all the outputs into a batch specific (i.e. camera name and model) directory.
+        mkdir --parents "/data/kalibr/${camera_name}"
+        mv -- /data/*.pdf /data/*.txt /data/*.yaml "/data/kalibr/${camera_name}"
+
+    done < <(jq ".cameras.[]" "${dataset_specification_json}")
+done < <(jq ".bags.[]" "${dataset_specification_json}")
